Split CloudFormation functions out into separate file
